_legacy_forge/underwriteos/pyproject.toml
[project]
name = "underwriteos"
version = "0.1.0"
description = "Bank-grade SBA underwriting memo engine"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=4.0",
    "pdfplumber>=0.11",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.5", "mypy>=1.10"]

[project.scripts]
underwriteos = "underwriteos.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
